ci: Add dummy workflow to build images
Add a dummy workflow so the actual workflow can be developed. This is
standard procedure as you cannot dispatch to a workflow that does not
exist on the main branch, even if a dev branch has the workflow.

A subsequent PR will flesh out the workflow to actually build the
container images and push to ECR.