Sanitize version/ref_name for use as docker label
Docker labels/tags can contain only alphanumeric chars, underscores,
periods and hyphens and must start with alphanumeric or underscores. So
we convert all invalid chars to hyphens and strip leading invalid chars.
Also collapse runs of hyphens to a single hyphen.

Technically we do not need to strip the leading chars as we construct
the label with the version as a suffix, but it would look a little weird
if we didn't and allows us to use the version as a pure tag if we want
to later.